src/requirements.txt
flask
openai
llamaapi
python-dotenv
seaborn
graphviz
matplotlib
numpy
pandas